pyproject.toml
[tool.poetry]
name = "planetmint"
version = "2.3.0"
description = "Planetmint: The Blockchain Database"
authors = ["Planetmint contributors"]
license = "AGPLv3"
readme = "README.md"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Topic :: Database",
  "Topic :: Database :: Database Engines/Servers",
  "Topic :: Software Development",
  "Natural Language :: English",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.9",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: POSIX :: Linux"
]

[tool.poetry.scripts]
planetmint = "planetmint.commands.planetmint:main"

[tool.poetry.dependencies]
python = "^3.9"
chardet = "3.0.4"
base58 = "2.1.1"
aiohttp = "3.8.1"
flask-cors = "3.0.10"
flask-restful = "0.3.9"
flask = "2.1.2"
gunicorn = "20.1.0"
jsonschema = "4.16.0"
logstats = "0.3.0"
packaging = ">=22.0"
pymongo = "3.11.4"
tarantool = ">=0.12.1"
python-rapidjson = ">=1.0"
pyyaml = "6.0.0"
requests = "2.25.1"
setproctitle = "1.2.2"
werkzeug = "2.0.3"
nest-asyncio = "1.5.5"
protobuf = "3.20.2"
planetmint-ipld = ">=0.0.3"
pyasn1 = ">=0.4.8"
python-decouple = "^3.7"
planetmint-transactions = ">=0.7.0"
asynctnt = "^2.0.1"
abci = "^0.8.3"

[tool.poetry.group.dev.dependencies]
aafigure = "0.6"
alabaster = "0.7.12"
babel = "2.10.1"
certifi = "2022.12.7"
charset-normalizer = "2.0.12"
commonmark = "0.9.1"
docutils = "0.17.1"
idna = "2.10"
imagesize = "1.3.0"
importlib-metadata = "4.11.3"
jinja2 = "3.0.0"
markdown-it-py = "2.2.0"
markupsafe = "2.1.1"
mdit-py-plugins = "0.3.0"
mdurl = "0.1.1"
myst-parser = "0.17.2"
pockets = "0.9.1"
pygments = "2.12.0"
pyparsing = "3.0.8"
pytz = "2022.1"
pyyaml = ">=5.4.0"
six = "1.16.0"
snowballstemmer = "2.2.0"
sphinx = "4.5.0"
sphinx-rtd-theme = "1.0.0"
sphinxcontrib-applehelp = "1.0.2"
sphinxcontrib-devhelp = "1.0.2"
sphinxcontrib-htmlhelp = "2.0.0"
sphinxcontrib-httpdomain = "1.8.0"
sphinxcontrib-jsmath = "1.0.1"
sphinxcontrib-napoleon = "0.7"
sphinxcontrib-qthelp = "1.0.3"
sphinxcontrib-serializinghtml = "1.1.5"
urllib3 = "1.26.9"
wget = "3.2"
zipp = "3.8.0"
nest-asyncio = "1.5.5"
sphinx-press-theme = "0.8.0"
sphinx-documatt-theme = "^0.0.5"
ipdb = "^0.13.11"
ipython = "^8.9.0"
watchdog = "^2.2.1"
logging_tree = "^1.9"
pre-commit = "^3.0.3"
twine = "^4.0.2"
ptvsd = "^4.3.2"
coverage = "^7.1.0"
pep8 = "^1.7.1"
black = ">=23.1.0"
hypothesis = ">=5.3.0"
pytest = ">=3.0.0"
pytest-cov = "2.8.1"
pytest-mock = "^3.10.0"
pytest-xdist = "^3.1.0"
pytest-flask = "^1.2.0"
pytest-aiohttp = "^1.0.4"
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
